% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ideniforj}
\alias{ideniforj}
\title{Identify subject index from the validation set}
\usage{
ideniforj(Xi, Xj)
}
\arguments{
\item{Xi}{Vector, a vector of numeric values}

\item{Xj}{Vector, a vector of numeric values}
}
\value{
Returns a vector identifying valid \code{Xi} for \code{Xj}
}
\description{
Identify subject index from the validation set
}
\keyword{internal}
