% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idensigmaforj}
\alias{idensigmaforj}
\title{Determine variance-covariance matrix (with correct dimension)}
\usage{
idensigmaforj(Xi, Sigma)
}
\arguments{
\item{Xi}{Vector, a vector of numeric values}

\item{Sigma}{Matrix, a variance-covariance matrix}
}
\value{
Returns a variance-covariance matrix of correct dimension based on the index of \code{Xi}
}
\description{
Determine variance-covariance matrix (with correct dimension)
}
\keyword{internal}
