% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logdmvnorm}
\alias{logdmvnorm}
\title{Derive the logarithm of the likelihood of a multivariate normal distribution}
\usage{
logdmvnorm(X, mean, Sigma)
}
\arguments{
\item{X}{Vector, a vector of numeric values}

\item{mean}{Vector, the mean of multivariate normal}

\item{Sigma}{Matrix, the variance-covariance matrix of multivariate normal}
}
\value{
Returns the logarithm of the likelihood of X with multivariate normal with mean \code{mean} and variance \code{Sigma}
}
\description{
Derive the logarithm of the likelihood of a multivariate normal distribution
}
\keyword{internal}
