% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logliknonvalidXvaryauxinv}
\alias{logliknonvalidXvaryauxinv}
\title{Calculate the log-likelihood of the non-validation set (time-varying X and time-invariant auxiliary variables)}
\usage{
logliknonvalidXvaryauxinv(
  Yval,
  Ynonval,
  Xval,
  Xnonval,
  Z,
  auxval,
  auxnonval,
  theta,
  Sigma,
  H
)
}
\arguments{
\item{Yval}{Matrix, a matrix of outcomes from the validation set}

\item{Ynonval}{Matrix, a matrix of outcomes from the non-validation set}

\item{Xval}{Matrix, a matrix of missing covariates from the validation set}

\item{Xnonval}{Matrix, a matrix of missing covariates from the non-validation set}

\item{Z}{Matrix, a matrix with fully-observed covariates}

\item{auxval}{Matrix, a matrix of auxiliary variables from the validation set}

\item{auxnonval}{Matrix, a matrix of auxiliary variables from the non-validation set}

\item{theta}{Vector, a vector of unknown parameters}

\item{Sigma}{Matrix, a matrix of variance-covariance structure}

\item{H}{Vector, a vector of bandwidths}
}
\value{
Returns the log-likelihood contributed by the non-validation set with time-varying X and time-invariant auxliary variables
}
\description{
Calculate the log-likelihood of the non-validation set (time-varying X and time-invariant auxiliary variables)
}
\keyword{internal}
