% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logliknonvalidXvaryauxvary}
\alias{logliknonvalidXvaryauxvary}
\title{Calculate the log-likelihood of the non-validation set (time-varying X and time-varying auxiliary variables)}
\usage{
logliknonvalidXvaryauxvary(
  Yval,
  Ynonval,
  Xval,
  Xnonval,
  Z,
  auxval,
  auxnonval,
  theta,
  Sigma,
  H,
  auxcont
)
}
\arguments{
\item{Yval}{Matrix, a matrix of outcomes from the validation set}

\item{Ynonval}{Matrix, a matrix of outcomes from the non-validation set}

\item{Xval}{Matrix, a matrix of missing covariates from the validation set}

\item{Xnonval}{Matrix, a matrix of missing covariates from the non-validation set}

\item{Z}{Matrix, a matrix with fully-observed covariates}

\item{auxval}{Matrix, a matrix of auxiliary variables from the validation set}

\item{auxnonval}{Matrix, a matrix of auxiliary variables from the non-validation set}

\item{theta}{Vector, a vector of unknown parameters}

\item{Sigma}{Matrix, a matrix of variance-covariance structure}

\item{H}{Vector, a vector of bandwidths}

\item{auxcont, }{a logistic argument determining whether the auxiliary variables are continuous}
}
\value{
Returns the log-likelihood contributed by the non-validation set with time-varying X and time-varying auxliary variables
}
\description{
Calculate the log-likelihood of the non-validation set (time-varying X and time-varying auxiliary variables)
}
\keyword{internal}
