% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglikvalid}
\alias{loglikvalid}
\title{Calculate the log-likelhood of the validation set}
\usage{
loglikvalid(Y, X, Z, theta, Sigma)
}
\arguments{
\item{Y}{Matrix, a matrix of outcomes}

\item{X}{Matrix, a matrix with missing covariates}

\item{Z}{Matrix, a matrix with fully-observed covariates}

\item{theta}{Vector, a vector of unknown parameters}

\item{Sigma}{Matrix, a matrix of variance-covariance structure}
}
\value{
Returns the log-likelihood of the validation set
}
\description{
Calculate the log-likelhood of the validation set
}
\keyword{internal}
