% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_timevarying_S_timeinv.R
\name{nonpara_X_timevarying_S_timeinv}
\alias{nonpara_X_timevarying_S_timeinv}
\title{This function performs a nonparametric estimation with missing data in
time-varying covariates using time-varying auxilirary information.}
\usage{
nonpara_X_timevarying_S_timeinv(
  n.sub,
  m.visit,
  data.y,
  data.x,
  data.z,
  data.aux,
  para.ini,
  cov.cont = TRUE,
  aux.cont = TRUE
)
}
\arguments{
\item{n.sub}{is a numeric input of the total number of subjects}

\item{m.visit}{is a numeric input of the total (maximum) number of visits for
each subject}

\item{data.y}{is a matrix/data frame of all longitudinal outcomes.}

\item{data.x}{is a matrix/data frame of time-varying covariates containing
missing values.}

\item{data.z}{is a matrix/data frame of fully observed covariates.}

\item{data.aux}{is a matrix/data frame of auxiliary variables.}

\item{para.ini}{is a vector of initial values of parameters.}

\item{cov.cont}{is a logistic argument given by \code{TRUE} if the missing
covariates are continuous; \code{FALSE}, otherwise.}

\item{aux.cont}{is a logistic argument given by \code{TRUE} if the auxiliary
variable is continuous; \code{FALSE}, otherwise.}
}
\value{
a list including parameter estimates, standard errors and the
convergence code
}
\description{
This function performs a nonparametric estimation with missing data in
time-varying covariates using time-varying auxilirary information.
}
\note{
To be added
}
\references{
To be added
}
\keyword{auxiliary}
\keyword{covariates,}
\keyword{time-varing}
\keyword{time-varying}
\keyword{variables}
