% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparlongdat-package.R
\docType{package}
\name{nonparlongdat-package}
\alias{nonparlongdat}
\alias{nonparlongdat-package}
\title{nonparlongdat: Nonparametric Methods for Longitudinal Data Analysis}
\description{
This package provides functions to implement nonparametric methods for
longitudinal data analysis
\itemize{
\item TBA
\item TBA
\item TBA
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/panpanzhang99299/nonparlongdat}
  \item Report bugs at \url{https://github.com/panpanzhang99299/nonparlongdat/issues}
}

}
\author{
\strong{Maintainer}: Panpan Zhang \email{panpan.zhang@vumc.org} (\href{https://orcid.org/0000-0002-8211-5930}{ORCID})

Authors:
\itemize{
  \item Leah Suttner
  \item Sharon Xie \email{sxie@pennmedicine.upenn.edu} (\href{https://orcid.org/0000-0002-7886-4103}{ORCID})
}

}
\keyword{internal}
